WIP: change the semantics of `MatchesAndTheirSections.matchFrom` (and rename it to `MatchesAndTheirSections.allSidesMatchOf` so that never pares down all-sides matches to pairwise matches, and provisionally allows subsumption and just one-side.
This breaks things down stream, but we do go past the first failure in the bug described in https://github.com/sageserpent-open/kineticMerge/issues/91.

The next failure is an assertion in `MatchesAndTheirSections.eatIntoSection` that has been loosened after confirmation that the supposed failure was indeed valid.

Now we have a failure due to stack overflow when merging, plus various broken tests due to the downstream processing being affected by the new semantics...